simplify brev-metadata call
